+++ CHANGELOG.md
@@ -4,7 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+ - Definitions of the available properties as collapsible sections, typeset with MathJax.
+
+### Changed
+ - New introduction text with links to the related articles.
+
+### Fixed
+ - Property `skeleton-chordal` renamed to `skeleton-weakly-chordal` to match the
+   choice accepted by the detector script.
+
 ## [0.1.0] - 2023-03-28
+
 ### Fixed
  - Invalid meta viewport